fastapi/requirements.txt
fastapi==0.85.1
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.4
Pillow==9.2.0
pydantic==1.9.1
pytest==7.1.2
scikit_learn==1.1.2
scipy==1.7.3
xarray==0.20.1
